+++ src/agent_blueprint/models/__init__.py
@@ -2,5 +2,6 @@
 
 from agent_blueprint.models.blueprint import BlueprintSpec
 from agent_blueprint.models.contracts import ContractsDef
+from agent_blueprint.models.policies import PoliciesDef
 
-__all__ = ["BlueprintSpec", "ContractsDef"]
+__all__ = ["BlueprintSpec", "ContractsDef", "PoliciesDef"]